+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/crawler/factory/TanaguruCrawlerControllerFactory.java
@@ -25,6 +25,6 @@ public interface TanaguruCrawlerControllerFactory {
                                                int maxPage,
                                                int maxDepth,
                                                String basicAuthUrl,
-                                               String basicAuthPassword,
-                                               String basicAuthLogin) throws Exception;
+                                               String basicAuthLogin,
+                                               String basicAuthPassword) throws Exception;
 }