+++ CHANGELOG.md
@@ -55,3 +55,9 @@
 - New feature : SSO
 - Update project_user authorities : add ressources
 - update tests to 5.1.1
+
+## 1.0.7
+- Fixes:
+  - basic authentication on all type of audit
+  - remove project with apiKey
+- add getAuditsByAuditScheduler && getLastAuditByScheduledByAndAuditStatus audit endpoints